Move types out of AttributeSelector component to shared types
